+++ test/runtests.jl
@@ -33,7 +33,6 @@ end
     include("causality_tests/test_scalarseries_ApproximateSimplexIntersectionTest.jl")
     include("causality_tests/test_scalarseries_ExactSimplexIntersectionTest.jl")
     include("causality_tests/test_scalarseries_PredictiveAsymmetryTest.jl")
-    
     include("causality_tests/test_scalarseries_SMeasure.jl")
 end